login sql

+++ wechat/login.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `user` (
-  `id` bigint(64) unsigned NOT NULL AUTO_INCREMENT,
+  `id` int(20) unsigned NOT NULL AUTO_INCREMENT,
   `open_id` varchar(32) NOT NULL COMMENT '用户 open_id',
   `created_at` timestamp NULL DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间',
   `updated_at` timestamp NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT '更新时间',
